Fix render: f32 shaders, compute→vertex barrier, no stale uploads
Three render bugs fixed:

1. Vertex shader read float64_t but buffers are f32 — particles showed
   as garbage point at center. Fixed render_points.vert to use float.

2. No pipeline barrier between compute writes and vertex shader reads —
   render saw stale buffer data. Added VK_PIPELINE_STAGE_COMPUTE →
   VERTEX barrier in render command buffer.

3. Codegen uploaded stale CPU particle arrays OVER GPU results every
   frame ("Sync CPU state for render"). GPU computed correct positions,
   then codegen overwrote them with init data. Fixed: don't upload
   arrays that GPU kernels just wrote (gpu_written set subtracted
   from upload_needed).

4. frame_end must come BEFORE render_points, not after. Without CPU
   downloads triggering early frame_end, compute command buffer stayed
   open during render — render read uncommitted data.

Frame loop now: frame_begin → dispatches → frame_end → frame_wait →
render_points → should_close. Zero uploads, zero downloads.

+++ mcl/runtime/render_points.vert
@@ -1,20 +1,20 @@
 #version 450
-#extension GL_EXT_shader_explicit_arithmetic_types_float64 : enable
 
 /*
  * Point cloud vertex shader for particle simulations.
  *
  * Each invocation = one particle. gl_VertexIndex = particle index.
- * Reads position from 3 separate f64 storage buffers (SoA layout).
- * Reads color value from a 4th f64 buffer.
+ * Reads position from 3 separate storage buffers (SoA layout).
+ * Reads color value from a 4th buffer.
+ * Supports f32 (--precision f32) via float buffers.
  */
 
 layout(location = 0) out float v_value; /* normalized color value */
 
-layout(set = 0, binding = 0) buffer BufX { float64_t px[]; };
-layout(set = 0, binding = 1) buffer BufY { float64_t py[]; };
-layout(set = 0, binding = 2) buffer BufZ { float64_t pz[]; };
-layout(set = 0, binding = 3) buffer BufC { float64_t cv[]; };
+layout(set = 0, binding = 0) buffer BufX { float px[]; };
+layout(set = 0, binding = 1) buffer BufY { float py[]; };
+layout(set = 0, binding = 2) buffer BufZ { float pz[]; };
+layout(set = 0, binding = 3) buffer BufC { float cv[]; };
 
 layout(push_constant) uniform PC {
     mat4  viewProj;
@@ -27,14 +27,14 @@ layout(push_constant) uniform PC {
 void main() {
     int i = gl_VertexIndex;
 
-    float x = float(px[i]) * pc.world_scale;
-    float y = float(py[i]) * pc.world_scale;
-    float z = float(pz[i]) * pc.world_scale;
+    float x = px[i] * pc.world_scale;
+    float y = py[i] * pc.world_scale;
+    float z = pz[i] * pc.world_scale;
 
     gl_Position = pc.viewProj * vec4(x, y, z, 1.0);
     gl_PointSize = pc.point_size;
 
-    float val = float(cv[i]);
+    float val = cv[i];
     v_value = clamp((val - pc.val_min) / (pc.val_max - pc.val_min + 1e-10),
                     0.0, 1.0);
 }
